requirements.txt
opencv-python
numpy
mediapipe
ultralytics
google-generativeai # or openai
PyQt6
Pillow
python-dotenv
# If using local LLaVA (advanced):
# torch
# transformers
# peft
# bitsandbytes
# accelerate
# sentencepiece
# einops